Move color resources to Helpers group
