+++ MovieQuiz/Helpers/UIColor+Extensions.swift
@@ -1,4 +1,11 @@
 import Foundation
 import UIKit
 
-extension UIColor { }
+extension UIColor {
+    static var ypBlack: UIColor { UIColor(named: "YP Black") ?? .black }
+    static var ypWhite: UIColor { UIColor(named: "YP White") ?? .white }
+    static var ypGreen: UIColor { UIColor(named: "YP Green") ?? .green }
+    static var ypRed: UIColor { UIColor(named: "YP Red") ?? .red }
+    static var ypGray: UIColor { UIColor(named: "YP Gray") ?? .gray }
+    static var ypBackground: UIColor { UIColor(named: "YP Background") ?? .systemBackground }
+}
