--- MovieQuiz/Resources/Colors.swift
@@ -1,17 +1,0 @@
-//
-//  Colors.swift
-//  MovieQuiz
-//
-//  Created by Andrei Chenchik on 27/7/22.
-//
-
-import UIKit
-
-extension UIColor {
-    static var ypBlack: UIColor { UIColor(named: "YP Black") ?? .black }
-    static var ypWhite: UIColor { UIColor(named: "YP White") ?? .white }
-    static var ypGreen: UIColor { UIColor(named: "YP Green") ?? .green }
-    static var ypRed: UIColor { UIColor(named: "YP Red") ?? .red }
-    static var ypGray: UIColor { UIColor(named: "YP Gray") ?? .gray }
-    static var ypBackground: UIColor { UIColor(named: "YP Background") ?? .systemBackground }
-}
